refactor(web): extract cloneAndNormalize and blank-fixture helpers
